Feat: convert all money calculations to BigDecimal type

+++ src/main/java/com/counter/budget/budgetcounterbe/dto/CreateTransactionRequest.java
@@ -3,11 +3,12 @@ package com.counter.budget.budgetcounterbe.dto;
 import com.counter.budget.budgetcounterbe.model.TransactionType;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.UUID;
 
 public record CreateTransactionRequest(
-        @NotNull float amount,
+        @NotNull BigDecimal amount,
         @NotNull TransactionType type,
         UUID bucketId
 ) {}
